bugfix/error_signing_redirect: Added error handling
* Added a retrofit exception that transforms all the retrofit rx calls
to the retrofit exception. Retrofit exception is of 3 types, http,
network and unexpected.
* Added WebViewException, RetrofitException and EmptyResultsException.
WebViewException is called when the webview fails to load an url.
EmptyResultsException is thrown in case the data retrieved from both
database and network is null or empty.
* Added a base fragment, namely, BaseViewModelFragment where we handle
the error received from the viewmodel by overriding onError method.
* Added error handling when login fails.
* Replaced all this from methods like viewModel.obj.observer(this, ...)
with viewLifecycleOwner.
* Created a few utility methods in Utils.
* Improvd Result class so now it supports success and error checks.
* Added error codes for not authorized, unknown and not found.
* Added some error messages to the strings.xml.
* Added logs for unknown actions.
* Added todos for not fixed errors.